+++ src/lib.rs
@@ -12,11 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Pushrod is a Cross Platform UI Widget Library for Piston.
+//! Pushrod is a Cross Platform UI Widget Library for Rust.
 //!
 //! It is intended to be lightweight, easy to use, and easy to understand.  Pushrod draws
-//! inspiration from Atari ST GEM development, TrollTech Qt, and other libraries over
-//! the years.
+//! inspiration from 16-bit GUI-based systems and other libraries over the years.
 //!
 //! Pushrod uses Piston as its main window drawing and event loop functionality.  It utilizes
 //! piston2d-opengl graphics so that - eventually - the graphics can be represented as 3D
@@ -26,10 +25,8 @@
 //! Pushrod uses the following dependencies:
 //! ```ignore
 //! [dependencies]
-//! piston_window = "^0.87.0"
-//! piston2d-opengl_graphics = "^0.59.0"
-//! gfx_core = "^0.8.3"
-//! gfx_device_gl = "^0.15.5"
+//! piston_window = "^0.89.0"
+//! find_folder = "^0.3.0"
 //! ```
 
 /// Main module containing the run loop for the UI components, containers for windows and
@@ -49,4 +46,6 @@ pub mod event;
 /// Currently contains:
 /// - Base Widget (for drawing a plain background)
 /// - Box Widget (for drawing a plain background with a box and a colored border)
+/// - Text Widget (for drawing text)
+/// - Timer Widget (for performing timer operations)
 pub mod widget;