+++ RELEASE_NOTES.md
@@ -1,5 +1,20 @@
 # Pushrod Releases
 
+## 0.1.15
+
+- Made a new `BlankCallback` type for timer
+- Changed timer code to use the `BlankCallback` in the CallbackStore.
+- Optimized Cargo.toml file to only include one dependency: piston.
+- Modified run loop to include window factory object retrieval.
+- Added Image widget, added official 512x512 logo.
+- Added Image auto-scaling when drawing.
+- Fixed text widget so it uses the text color when drawing text.
+- Currently, text is causing issues with image display.
+
+## 0.1.14
+
+- Interim release to fix bugs in build and tests
+
 ## 0.1.13
 
 - Removed texture creation and renewal code.